Time: 204 ms (74.62%) | Memory: 10.8 MB (52.65%) - LeetSync

+++ 3422-find-the-n-th-value-after-k-seconds/find-the-n-th-value-after-k-seconds.cpp
@@ -1,7 +1,7 @@
 class Solution {
 public:
     int valueAfterKSeconds(int n, int k) {
-        vector<long long> pref(n,1);
+        vector<int> pref(n,1);
         long long mod = 1000000007;
 
         while(k--){
